+++ client/src/App.css
@@ -243,3 +243,25 @@ body {
       background-color: #d478df;
   }
 }
+
+
+.task-list-container {
+  max-height: 40vh; /* Adjust as needed */
+  overflow-y: auto;
+  border-radius: 5px;
+  padding: 10px;
+}
+
+/* Optional: Customize scrollbar appearance */
+.task-list-container::-webkit-scrollbar {
+  width: 6px;
+}
+
+.task-list-container::-webkit-scrollbar-thumb {
+  background-color: #d69cdd;
+  border-radius: 10px;
+}
+
+.task-list-container::-webkit-scrollbar-thumb:hover {
+  background-color: #d478df;
+}
